Updated Program
Made the window always be the top window and also lock resizing of the window so that when the user brings it into the live streaming program and does a window crop, resizing the window will not mess it up. Also fixed a bug where the scoreboard time was lagging a second behind.